% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{mlr3verse_deps}
\alias{mlr3verse_deps}
\title{List all mlr3verse dependencies}
\usage{
mlr3verse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
mlr3verse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all mlr3verse dependencies
}
