% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3verse_info.R
\name{mlr3verse_info}
\alias{mlr3verse_info}
\title{Print package versions}
\usage{
mlr3verse_info()
}
\description{
Returns the package versions of loaded mlr3 packages as a \link{data.frame}.
}
\examples{
mlr3verse_info()
}
