% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{mlr3verse_update}
\alias{mlr3verse_update}
\title{Update mlr3verse packages}
\usage{
mlr3verse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
mlr3verse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all mlr3verse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
mlr3verse_update()
}
}
