% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_mlr3misc.R, R/reexports_bbotk.R,
%   R/reexports_data.table.R, R/reexports_mlr3cluster.R,
%   R/reexports_mlr3filters.R, R/reexports_mlr3fselect.R,
%   R/reexports_mlr3hyperband.R, R/reexports_mlr3pipelines.R,
%   R/reexports_mlr3torch.R, R/reexports_mlr3tuning.R,
%   R/reexports_mlr3tuningspaces.R, R/reexports_mlr3viz.R,
%   R/reexports_paradox.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{clbk}
\alias{clbks}
\alias{mlr_callbacks}
\alias{mlr_terminators}
\alias{trm}
\alias{trms}
\alias{as.data.table}
\alias{as_task_clust}
\alias{as_prediction_clust}
\alias{mlr_filters}
\alias{flt}
\alias{flts}
\alias{auto_fselector}
\alias{callback_batch_fselect}
\alias{ensemble_fselect}
\alias{extract_inner_fselect_results}
\alias{extract_inner_fselect_archives}
\alias{fselect_nested}
\alias{fselect}
\alias{mlr_fselectors}
\alias{fs}
\alias{fss}
\alias{fsi}
\alias{hyperband_budget}
\alias{hyperband_n_configs}
\alias{hyperband_schedule}
\alias{mlr_pipeops}
\alias{po}
\alias{pos}
\alias{mlr_graphs}
\alias{ppl}
\alias{\%>>\%}
\alias{\%>>!\%}
\alias{gunion}
\alias{chain_graphs}
\alias{concat_graphs}
\alias{GraphLearner}
\alias{as_graph}
\alias{as_pipeop}
\alias{selector_all}
\alias{selector_cardinality_greater_than}
\alias{selector_grep}
\alias{selector_intersect}
\alias{selector_invert}
\alias{selector_missing}
\alias{selector_name}
\alias{selector_none}
\alias{selector_setdiff}
\alias{selector_type}
\alias{selector_union}
\alias{mlr3torch_callbacks}
\alias{t_clbk}
\alias{t_clbks}
\alias{mlr3torch_losses}
\alias{t_loss}
\alias{t_losses}
\alias{mlr3torch_optimizers}
\alias{t_opt}
\alias{t_opts}
\alias{nn}
\alias{as_torch_callback}
\alias{as_torch_callbacks}
\alias{as_torch_loss}
\alias{as_torch_optimizer}
\alias{lazy_tensor}
\alias{as_lazy_tensor}
\alias{materialize}
\alias{auto_tuner}
\alias{callback_async_tuning}
\alias{callback_batch_tuning}
\alias{extract_inner_tuning_results}
\alias{extract_inner_tuning_archives}
\alias{mlr_tuners}
\alias{tnr}
\alias{tnrs}
\alias{ti}
\alias{ti_async}
\alias{tune_nested}
\alias{tune}
\alias{mlr_tuning_spaces}
\alias{lts}
\alias{ltss}
\alias{autoplot}
\alias{fortify}
\alias{plot_learner_prediction}
\alias{ps}
\alias{p_lgl}
\alias{p_int}
\alias{p_dbl}
\alias{p_fct}
\alias{p_uty}
\alias{to_tune}
\alias{generate_design_grid}
\alias{generate_design_lhs}
\alias{generate_design_random}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bbotk}{\code{\link[bbotk:mlr_terminators]{mlr_terminators}}, \code{\link[bbotk:trm]{trm()}}, \code{\link[bbotk:trms]{trms()}}}

  \item{data.table}{\code{\link[data.table:as.data.table]{as.data.table()}}}

  \item{mlr3cluster}{\code{\link[mlr3cluster:as_prediction_clust]{as_prediction_clust()}}, \code{\link[mlr3cluster:as_task_clust]{as_task_clust()}}}

  \item{mlr3filters}{\code{\link[mlr3filters:flt]{flt()}}, \code{\link[mlr3filters:flts]{flts()}}, \code{\link[mlr3filters:mlr_filters]{mlr_filters}}}

  \item{mlr3fselect}{\code{\link[mlr3fselect:auto_fselector]{auto_fselector()}}, \code{\link[mlr3fselect:callback_batch_fselect]{callback_batch_fselect()}}, \code{\link[mlr3fselect:ensemble_fselect]{ensemble_fselect()}}, \code{\link[mlr3fselect:extract_inner_fselect_archives]{extract_inner_fselect_archives()}}, \code{\link[mlr3fselect:extract_inner_fselect_results]{extract_inner_fselect_results()}}, \code{\link[mlr3fselect:fs]{fs()}}, \code{\link[mlr3fselect:fselect]{fselect()}}, \code{\link[mlr3fselect:fselect_nested]{fselect_nested()}}, \code{\link[mlr3fselect:fsi]{fsi()}}, \code{\link[mlr3fselect:fss]{fss()}}, \code{\link[mlr3fselect:mlr_fselectors]{mlr_fselectors}}}

  \item{mlr3hyperband}{\code{\link[mlr3hyperband:hyperband_budget]{hyperband_budget()}}, \code{\link[mlr3hyperband:hyperband_n_configs]{hyperband_n_configs()}}, \code{\link[mlr3hyperband:hyperband_schedule]{hyperband_schedule()}}}

  \item{mlr3misc}{\code{\link[mlr3misc:clbk]{clbk()}}, \code{\link[mlr3misc:clbks]{clbks()}}, \code{\link[mlr3misc:mlr_callbacks]{mlr_callbacks}}}

  \item{mlr3pipelines}{\code{\link[mlr3pipelines:\%>>!\%]{\%>>!\%}}, \code{\link[mlr3pipelines:\%>>\%]{\%>>\%}}, \code{\link[mlr3pipelines:as_graph]{as_graph()}}, \code{\link[mlr3pipelines:as_pipeop]{as_pipeop()}}, \code{\link[mlr3pipelines:chain_graphs]{chain_graphs()}}, \code{\link[mlr3pipelines:concat_graphs]{concat_graphs()}}, \code{\link[mlr3pipelines:GraphLearner]{GraphLearner}}, \code{\link[mlr3pipelines:gunion]{gunion()}}, \code{\link[mlr3pipelines:mlr_graphs]{mlr_graphs}}, \code{\link[mlr3pipelines:mlr_pipeops]{mlr_pipeops}}, \code{\link[mlr3pipelines:po]{po()}}, \code{\link[mlr3pipelines:pos]{pos()}}, \code{\link[mlr3pipelines:ppl]{ppl()}}, \code{\link[mlr3pipelines:selector_all]{selector_all()}}, \code{\link[mlr3pipelines:selector_cardinality_greater_than]{selector_cardinality_greater_than()}}, \code{\link[mlr3pipelines:selector_grep]{selector_grep()}}, \code{\link[mlr3pipelines:selector_intersect]{selector_intersect()}}, \code{\link[mlr3pipelines:selector_invert]{selector_invert()}}, \code{\link[mlr3pipelines:selector_missing]{selector_missing()}}, \code{\link[mlr3pipelines:selector_name]{selector_name()}}, \code{\link[mlr3pipelines:selector_none]{selector_none()}}, \code{\link[mlr3pipelines:selector_setdiff]{selector_setdiff()}}, \code{\link[mlr3pipelines:selector_type]{selector_type()}}, \code{\link[mlr3pipelines:selector_union]{selector_union()}}}

  \item{mlr3torch}{\code{\link[mlr3torch:as_lazy_tensor]{as_lazy_tensor()}}, \code{\link[mlr3torch:as_torch_callback]{as_torch_callback()}}, \code{\link[mlr3torch:as_torch_callbacks]{as_torch_callbacks()}}, \code{\link[mlr3torch:as_torch_loss]{as_torch_loss()}}, \code{\link[mlr3torch:as_torch_optimizer]{as_torch_optimizer()}}, \code{\link[mlr3torch:lazy_tensor]{lazy_tensor()}}, \code{\link[mlr3torch:materialize]{materialize()}}, \code{\link[mlr3torch:mlr3torch_callbacks]{mlr3torch_callbacks}}, \code{\link[mlr3torch:mlr3torch_losses]{mlr3torch_losses}}, \code{\link[mlr3torch:mlr3torch_optimizers]{mlr3torch_optimizers}}, \code{\link[mlr3torch:nn]{nn()}}, \code{\link[mlr3torch:t_clbk]{t_clbk()}}, \code{\link[mlr3torch:t_clbks]{t_clbks()}}, \code{\link[mlr3torch:t_loss]{t_loss()}}, \code{\link[mlr3torch:t_losses]{t_losses()}}, \code{\link[mlr3torch:t_opt]{t_opt()}}, \code{\link[mlr3torch:t_opts]{t_opts()}}}

  \item{mlr3tuning}{\code{\link[mlr3tuning:auto_tuner]{auto_tuner()}}, \code{\link[mlr3tuning:callback_async_tuning]{callback_async_tuning()}}, \code{\link[mlr3tuning:callback_batch_tuning]{callback_batch_tuning()}}, \code{\link[mlr3tuning:extract_inner_tuning_archives]{extract_inner_tuning_archives()}}, \code{\link[mlr3tuning:extract_inner_tuning_results]{extract_inner_tuning_results()}}, \code{\link[mlr3tuning:mlr_tuners]{mlr_tuners}}, \code{\link[mlr3tuning:ti]{ti()}}, \code{\link[mlr3tuning:ti_async]{ti_async()}}, \code{\link[mlr3tuning:tnr]{tnr()}}, \code{\link[mlr3tuning:tnrs]{tnrs()}}, \code{\link[mlr3tuning:tune]{tune()}}, \code{\link[mlr3tuning:tune_nested]{tune_nested()}}}

  \item{mlr3tuningspaces}{\code{\link[mlr3tuningspaces:lts]{lts()}}, \code{\link[mlr3tuningspaces:ltss]{ltss()}}, \code{\link[mlr3tuningspaces:mlr_tuning_spaces]{mlr_tuning_spaces}}}

  \item{mlr3viz}{\code{\link[mlr3viz:autoplot]{autoplot}}, \code{\link[mlr3viz:fortify]{fortify}}, \code{\link[mlr3viz:plot_learner_prediction]{plot_learner_prediction()}}}

  \item{paradox}{\code{\link[paradox:generate_design_grid]{generate_design_grid()}}, \code{\link[paradox:generate_design_lhs]{generate_design_lhs()}}, \code{\link[paradox:generate_design_random]{generate_design_random()}}, \code{\link[paradox:p_dbl]{p_dbl()}}, \code{\link[paradox:p_fct]{p_fct()}}, \code{\link[paradox:p_int]{p_int()}}, \code{\link[paradox:p_lgl]{p_lgl()}}, \code{\link[paradox:p_uty]{p_uty()}}, \code{\link[paradox:ps]{ps()}}, \code{\link[paradox:to_tune]{to_tune()}}}
}}

