% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3verse-package}
\alias{mlr3verse}
\alias{mlr3verse-package}
\title{mlr3verse: Easily Install and Load the 'mlr3' Package Family}
\description{
The 'mlr3' package family is a set of packages for
    machine-learning purposes built in a modular fashion. This wrapper
    package is aimed to simplify the installation and loading of the core
    'mlr3' packages. Get more information about the 'mlr3' project at
    <https://mlr3book.mlr-org.com/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3verse.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3verse}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3verse/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
}

}
\keyword{internal}
