% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3verse_info.R
\name{mlr3verse_info}
\alias{mlr3verse_info}
\title{Print mlr3* package versions}
\usage{
mlr3verse_info()
}
\description{
Returns the package versions of imported and suggested mlr3 packages as a \code{\link[data.table:data.table]{data.table::data.table()}}.
Suggested packages that are not installed are listed with version \code{NA}.
}
\examples{
mlr3verse_info()
}
