% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{mlr3verse_update}
\alias{mlr3verse_update}
\title{Update mlr3verse packages}
\usage{
mlr3verse_update(recursive = FALSE, repos = getOption("repos"), type = "all")
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of mlr3verse
packages.}

\item{repos}{The repositories to use to check for updates. Defaults to
\code{getOptions("repos")}.}

\item{type}{Which package type to use. Valid options are "cran", "dev" and
"all".}
}
\description{
This will check to see if all mlr3verse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\section{Type}{

  When parameter `type = "all"`, all packages from the mlr3
  ecosystem will be updated using the CRAN versions for the ones which are on
  CRAN and the development versions from Github for those which are not yet
  released. When `type = "dev"` is set, all packages will be updated using
  their Github versions.
}

\examples{
\dontrun{
mlr3verse_update()
}
}
