% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports_bbotk.R, R/reexports_data.table.R,
%   R/reexports_mlr3.R, R/reexports_mlr3filters.R, R/reexports_mlr3fselect.R,
%   R/reexports_mlr3pipelines.R, R/reexports_mlr3tuning.R,
%   R/reexports_mlr3viz.R, R/reexports_paradox.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mlr_terminators}
\alias{trm}
\alias{trms}
\alias{as.data.table}
\alias{mlr_tasks}
\alias{tsk}
\alias{tsks}
\alias{mlr_task_generators}
\alias{tgen}
\alias{tgens}
\alias{TaskClassif}
\alias{TaskRegr}
\alias{mlr_learners}
\alias{lrn}
\alias{lrns}
\alias{mlr_resamplings}
\alias{rsmp}
\alias{rsmps}
\alias{mlr_measures}
\alias{msr}
\alias{msrs}
\alias{resample}
\alias{benchmark_grid}
\alias{benchmark}
\alias{as_data_backend}
\alias{as_benchmark_result}
\alias{mlr_filters}
\alias{flt}
\alias{flts}
\alias{mlr_fselectors}
\alias{fs}
\alias{mlr_pipeops}
\alias{po}
\alias{mlr_graphs}
\alias{ppl}
\alias{\%>>\%}
\alias{greplicate}
\alias{gunion}
\alias{GraphLearner}
\alias{as_graph}
\alias{selector_all}
\alias{selector_cardinality_greater_than}
\alias{selector_grep}
\alias{selector_intersect}
\alias{selector_invert}
\alias{selector_missing}
\alias{selector_name}
\alias{selector_none}
\alias{selector_setdiff}
\alias{selector_type}
\alias{selector_union}
\alias{mlr_tuners}
\alias{tnr}
\alias{tnrs}
\alias{AutoTuner}
\alias{TuningInstanceSingleCrit}
\alias{TuningInstanceMultiCrit}
\alias{autoplot}
\alias{fortify}
\alias{ParamSet}
\alias{ParamLgl}
\alias{ParamInt}
\alias{ParamDbl}
\alias{ParamFct}
\alias{generate_design_grid}
\alias{generate_design_lhs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bbotk}{\code{\link[bbotk]{mlr_terminators}}, \code{\link[bbotk]{trm}}, \code{\link[bbotk:trm]{trms}}}

  \item{data.table}{\code{\link[data.table]{as.data.table}}}

  \item{mlr3}{\code{\link[mlr3]{TaskClassif}}, \code{\link[mlr3]{TaskRegr}}, \code{\link[mlr3]{as_benchmark_result}}, \code{\link[mlr3]{as_data_backend}}, \code{\link[mlr3]{benchmark}}, \code{\link[mlr3]{benchmark_grid}}, \code{\link[mlr3:mlr_sugar]{lrn}}, \code{\link[mlr3:mlr_sugar]{lrns}}, \code{\link[mlr3]{mlr_learners}}, \code{\link[mlr3]{mlr_measures}}, \code{\link[mlr3]{mlr_resamplings}}, \code{\link[mlr3]{mlr_task_generators}}, \code{\link[mlr3]{mlr_tasks}}, \code{\link[mlr3:mlr_sugar]{msr}}, \code{\link[mlr3:mlr_sugar]{msrs}}, \code{\link[mlr3]{resample}}, \code{\link[mlr3:mlr_sugar]{rsmp}}, \code{\link[mlr3:mlr_sugar]{rsmps}}, \code{\link[mlr3:mlr_sugar]{tgen}}, \code{\link[mlr3:mlr_sugar]{tgens}}, \code{\link[mlr3:mlr_sugar]{tsk}}, \code{\link[mlr3:mlr_sugar]{tsks}}}

  \item{mlr3filters}{\code{\link[mlr3filters]{flt}}, \code{\link[mlr3filters:flt]{flts}}, \code{\link[mlr3filters]{mlr_filters}}}

  \item{mlr3fselect}{\code{\link[mlr3fselect]{fs}}, \code{\link[mlr3fselect]{mlr_fselectors}}}

  \item{mlr3pipelines}{\code{\link[mlr3pipelines:grapes-greater-than-greater-than-grapes]{\%>>\%}}, \code{\link[mlr3pipelines:mlr_learners_graph]{GraphLearner}}, \code{\link[mlr3pipelines]{as_graph}}, \code{\link[mlr3pipelines]{greplicate}}, \code{\link[mlr3pipelines]{gunion}}, \code{\link[mlr3pipelines]{mlr_graphs}}, \code{\link[mlr3pipelines]{mlr_pipeops}}, \code{\link[mlr3pipelines]{po}}, \code{\link[mlr3pipelines]{ppl}}, \code{\link[mlr3pipelines:Selector]{selector_all}}, \code{\link[mlr3pipelines:Selector]{selector_cardinality_greater_than}}, \code{\link[mlr3pipelines:Selector]{selector_grep}}, \code{\link[mlr3pipelines:Selector]{selector_intersect}}, \code{\link[mlr3pipelines:Selector]{selector_invert}}, \code{\link[mlr3pipelines:Selector]{selector_missing}}, \code{\link[mlr3pipelines:Selector]{selector_name}}, \code{\link[mlr3pipelines:Selector]{selector_none}}, \code{\link[mlr3pipelines:Selector]{selector_setdiff}}, \code{\link[mlr3pipelines:Selector]{selector_type}}, \code{\link[mlr3pipelines:Selector]{selector_union}}}

  \item{mlr3tuning}{\code{\link[mlr3tuning]{AutoTuner}}, \code{\link[mlr3tuning]{TuningInstanceMultiCrit}}, \code{\link[mlr3tuning]{TuningInstanceSingleCrit}}, \code{\link[mlr3tuning]{mlr_tuners}}, \code{\link[mlr3tuning]{tnr}}, \code{\link[mlr3tuning:tnr]{tnrs}}}

  \item{mlr3viz}{\code{\link[mlr3viz:reexports]{autoplot}}, \code{\link[mlr3viz:reexports]{fortify}}}

  \item{paradox}{\code{\link[paradox]{ParamDbl}}, \code{\link[paradox]{ParamFct}}, \code{\link[paradox]{ParamInt}}, \code{\link[paradox]{ParamLgl}}, \code{\link[paradox]{ParamSet}}, \code{\link[paradox]{generate_design_grid}}, \code{\link[paradox]{generate_design_lhs}}}
}}

