% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_mlr3misc.R, R/reexports_bbotk.R,
%   R/reexports_data.table.R, R/reexports_mlr3cluster.R,
%   R/reexports_mlr3filters.R, R/reexports_mlr3fselect.R,
%   R/reexports_mlr3pipelines.R, R/reexports_mlr3tuning.R,
%   R/reexports_mlr3tuningspaces.R, R/reexports_mlr3viz.R, R/reexports_paradox.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{clbk}
\alias{clbks}
\alias{mlr_terminators}
\alias{trm}
\alias{trms}
\alias{as.data.table}
\alias{TaskClust}
\alias{as_task_clust}
\alias{as_prediction_clust}
\alias{mlr_filters}
\alias{flt}
\alias{flts}
\alias{mlr_fselectors}
\alias{fs}
\alias{AutoFSelector}
\alias{FSelectInstanceSingleCrit}
\alias{FSelectInstanceMultiCrit}
\alias{extract_inner_fselect_results}
\alias{extract_inner_fselect_archives}
\alias{fselect}
\alias{fselect_nested}
\alias{auto_fselector}
\alias{fsi}
\alias{mlr_pipeops}
\alias{po}
\alias{pos}
\alias{mlr_graphs}
\alias{ppl}
\alias{\%>>\%}
\alias{\%>>!\%}
\alias{greplicate}
\alias{gunion}
\alias{chain_graphs}
\alias{concat_graphs}
\alias{GraphLearner}
\alias{as_graph}
\alias{as_pipeop}
\alias{selector_all}
\alias{selector_cardinality_greater_than}
\alias{selector_grep}
\alias{selector_intersect}
\alias{selector_invert}
\alias{selector_missing}
\alias{selector_name}
\alias{selector_none}
\alias{selector_setdiff}
\alias{selector_type}
\alias{selector_union}
\alias{mlr_tuners}
\alias{tnr}
\alias{tnrs}
\alias{AutoTuner}
\alias{TuningInstanceSingleCrit}
\alias{TuningInstanceMultiCrit}
\alias{extract_inner_tuning_results}
\alias{extract_inner_tuning_archives}
\alias{tune}
\alias{tune_nested}
\alias{auto_tuner}
\alias{ti}
\alias{mlr_tuning_spaces}
\alias{lts}
\alias{ltss}
\alias{autoplot}
\alias{fortify}
\alias{plot_learner_prediction}
\alias{ps}
\alias{p_lgl}
\alias{p_int}
\alias{p_dbl}
\alias{p_fct}
\alias{p_uty}
\alias{to_tune}
\alias{generate_design_grid}
\alias{generate_design_lhs}
\alias{generate_design_random}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bbotk}{\code{\link[bbotk]{mlr_terminators}}, \code{\link[bbotk]{trm}}, \code{\link[bbotk:trm]{trms}}}

  \item{data.table}{\code{\link[data.table]{as.data.table}}}

  \item{mlr3cluster}{\code{\link[mlr3cluster]{as_prediction_clust}}, \code{\link[mlr3cluster]{as_task_clust}}, \code{\link[mlr3cluster]{TaskClust}}}

  \item{mlr3filters}{\code{\link[mlr3filters]{flt}}, \code{\link[mlr3filters:flt]{flts}}, \code{\link[mlr3filters]{mlr_filters}}}

  \item{mlr3fselect}{\code{\link[mlr3fselect]{auto_fselector}}, \code{\link[mlr3fselect]{AutoFSelector}}, \code{\link[mlr3fselect]{extract_inner_fselect_archives}}, \code{\link[mlr3fselect]{extract_inner_fselect_results}}, \code{\link[mlr3fselect]{fs}}, \code{\link[mlr3fselect]{fselect}}, \code{\link[mlr3fselect]{fselect_nested}}, \code{\link[mlr3fselect]{FSelectInstanceMultiCrit}}, \code{\link[mlr3fselect]{FSelectInstanceSingleCrit}}, \code{\link[mlr3fselect]{fsi}}, \code{\link[mlr3fselect]{mlr_fselectors}}}

  \item{mlr3misc}{\code{\link[mlr3misc]{clbk}}, \code{\link[mlr3misc:clbk]{clbks}}}

  \item{mlr3pipelines}{\code{\link[mlr3pipelines:grapes-greater-than-greater-than-grapes]{\%>>!\%}}, \code{\link[mlr3pipelines:grapes-greater-than-greater-than-grapes]{\%>>\%}}, \code{\link[mlr3pipelines]{as_graph}}, \code{\link[mlr3pipelines]{as_pipeop}}, \code{\link[mlr3pipelines]{chain_graphs}}, \code{\link[mlr3pipelines:grapes-greater-than-greater-than-grapes]{concat_graphs}}, \code{\link[mlr3pipelines:mlr_learners_graph]{GraphLearner}}, \code{\link[mlr3pipelines]{greplicate}}, \code{\link[mlr3pipelines]{gunion}}, \code{\link[mlr3pipelines]{mlr_graphs}}, \code{\link[mlr3pipelines]{mlr_pipeops}}, \code{\link[mlr3pipelines]{po}}, \code{\link[mlr3pipelines:po]{pos}}, \code{\link[mlr3pipelines]{ppl}}, \code{\link[mlr3pipelines:Selector]{selector_all}}, \code{\link[mlr3pipelines:Selector]{selector_cardinality_greater_than}}, \code{\link[mlr3pipelines:Selector]{selector_grep}}, \code{\link[mlr3pipelines:Selector]{selector_intersect}}, \code{\link[mlr3pipelines:Selector]{selector_invert}}, \code{\link[mlr3pipelines:Selector]{selector_missing}}, \code{\link[mlr3pipelines:Selector]{selector_name}}, \code{\link[mlr3pipelines:Selector]{selector_none}}, \code{\link[mlr3pipelines:Selector]{selector_setdiff}}, \code{\link[mlr3pipelines:Selector]{selector_type}}, \code{\link[mlr3pipelines:Selector]{selector_union}}}

  \item{mlr3tuning}{\code{\link[mlr3tuning]{auto_tuner}}, \code{\link[mlr3tuning]{AutoTuner}}, \code{\link[mlr3tuning]{extract_inner_tuning_archives}}, \code{\link[mlr3tuning]{extract_inner_tuning_results}}, \code{\link[mlr3tuning]{mlr_tuners}}, \code{\link[mlr3tuning]{ti}}, \code{\link[mlr3tuning:mlr_sugar]{tnr}}, \code{\link[mlr3tuning:mlr_sugar]{tnrs}}, \code{\link[mlr3tuning]{tune}}, \code{\link[mlr3tuning]{tune_nested}}, \code{\link[mlr3tuning]{TuningInstanceMultiCrit}}, \code{\link[mlr3tuning]{TuningInstanceSingleCrit}}}

  \item{mlr3tuningspaces}{\code{\link[mlr3tuningspaces]{lts}}, \code{\link[mlr3tuningspaces:lts]{ltss}}, \code{\link[mlr3tuningspaces]{mlr_tuning_spaces}}}

  \item{mlr3viz}{\code{\link[mlr3viz:reexports]{autoplot}}, \code{\link[mlr3viz:reexports]{fortify}}, \code{\link[mlr3viz]{plot_learner_prediction}}}

  \item{paradox}{\code{\link[paradox]{generate_design_grid}}, \code{\link[paradox]{generate_design_lhs}}, \code{\link[paradox]{generate_design_random}}, \code{\link[paradox:Domain]{p_dbl}}, \code{\link[paradox:Domain]{p_fct}}, \code{\link[paradox:Domain]{p_int}}, \code{\link[paradox:Domain]{p_lgl}}, \code{\link[paradox:Domain]{p_uty}}, \code{\link[paradox]{ps}}, \code{\link[paradox]{to_tune}}}
}}

